% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_search_km_ge}
\alias{binary_search_km_ge}
\title{Binary search to find the location of the largest time that is smaller of the pattern}
\usage{
binary_search_km_ge(array, pattern)
}
\arguments{
\item{array}{A vector of sorted numbers to search the pattern in}

\item{pattern}{A number to be searched in the array}
}
\value{
The location of the largest time that is smaller of the pattern
}
\description{
Binary search to find the location of the largest time that is smaller of the pattern
}
\details{
This is an inner function inside KONPsurv package, not for the users.
}
