% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcinoma}
\alias{carcinoma}
\title{Urothelial carcinoma.}
\format{A data frame with 625 observations (316 in the atezolizumab group and 309 chemotherapy group) with the following 3 columns:
\describe{
  \item{time}{the observed follow-up times in days.}
  \item{status}{the event indicators, 0=right censored, 1= event.}
  \item{group}{ the group labels, 1 = atezolizumab, 2 = chemotherapy.}
}}
\usage{
data(carcinoma)
}
\description{
Survival data from a trial comparing chemotherapy versus atezolizumab in the treatment of Urothelial carcinoma.
}
\references{
Powles T, Dura?n I, van der Heijden MS, et al. Atezolizumab versus chemotherapy in patients with platinum-treated locally
advanced or metastatic urothelial carcinoma (IMvigor211): a multicentre, open-label, phase 3 randomised controlled trial.
Lancet 2018; 391: 748-757.
}
\keyword{datasets}
