% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_perm_stats}
\alias{get_perm_stats}
\title{A function to calculate the permutation test statistics}
\usage{
get_perm_stats(trt, ptrt_mat, time_original, delta_orginial,
  imputed_altern_time_vec, imputed_altern_delta_vec, n_perm)
}
\arguments{
\item{trt}{The original data Group vector, must contain 0 and 1 values only}

\item{ptrt_mat}{The permuted treatment matrix for all permutations}

\item{time_original}{The original data follow up time}

\item{delta_orginial}{The original data binary status vector, where 0 stands for censored observations and 1 stands for events}

\item{imputed_altern_time_vec}{The time vector for observations that switched groups in permuted sample}

\item{imputed_altern_delta_vec}{The delta vector for observations that switched groups in permuted sample}

\item{n_perm}{The number of permutations}
}
\value{
A list with the two permutations test stats vectors and a permutation table usage vector
}
\description{
A function to calculate the permutation test statistics
}
\details{
This is an inner function inside KONPsurv package, not for the users.
}
