% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_perm_stats_K_sample}
\alias{get_perm_stats_K_sample}
\title{A function to calculate the permutation test statistics for the KONP K-sample test}
\usage{
get_perm_stats_K_sample(ptrt_mat, imputed_time_matrix, imputed_delta_matrix,
  n_perm, n_vec)
}
\arguments{
\item{ptrt_mat}{The permuted treatment matrix for all permutations}

\item{imputed_time_matrix}{A matrix contianing the time each obsevations will recieve for each group it belongs}

\item{imputed_delta_matrix}{A matrix contianing the delta each obsevations will recieve for each group it belongs}

\item{n_perm}{The number of permutations}

\item{n_vec}{A vector with the sample sizes in each group}
}
\value{
A list with the two permutations test stats vectors
}
\description{
A function to calculate the permutation test statistics for the KONP K-sample test
}
\details{
This is an inner function inside SG package, not for the users.
}
