% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KM_C}
\alias{KM_C}
\title{An implementation of Kaplan Meier calculation in C++}
\usage{
KM_C(time, delta)
}
\arguments{
\item{time}{The follow up time}

\item{delta}{A binary status vector, where 0 stands for censored observations and 1 stands for events (corresponds with time)}
}
\value{
The function returns two vectors inside a list \cr

\code{time} - returns the sorted unique time of the data \cr 
\code{s} - returns the corresponding KM survival estimates to time
}
\description{
An implementation of Kaplan Meier calculation in C++
}
\details{
This is an inner function inside SG package, not for the users.
}
