% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_2_sample.R
\name{konp_2_sample}
\alias{konp_2_sample}
\title{KONP Test for Equality of Two Distributions for Right Censored Data}
\usage{
konp_2_sample(time, delta, trt, n.perm, n.impu = 1)
}
\arguments{
\item{time}{The follow up time.}

\item{delta}{A binary status vector, where 0 stands for censored observations and 1 stands for events.}

\item{trt}{Group vector, must contain only two unique values.}

\item{n.perm}{The number of permutations.}

\item{n.impu}{The number of imputations, for each imputation n.perm permutations will be executed.}
}
\value{
Two statistics and their appropriate P-values are returned: \cr 

\code{pv_chisq} - returns the P-value based on the Chi-square statistic. \cr 
\code{pv_lr} - returns the P-value based on the likelihood ratio statistic. \cr 
\code{chisq_test_stat} - returns the Chi-square test statistic. \cr 
\code{lr_test_stat} - returns the likelihood ratio test statistic.
}
\description{
KONP Test for Equality of Two Distributions for Right Censored Data
}
\details{
This is an inner function inside KONPSURV package, not for the users.
The user should run the K-sample test and if K=2 this function will run.
}
