% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_K_sample.R
\name{konp_test}
\alias{konp_test}
\title{KONP Tests for Equality of K Distributions for Right-Censored Data}
\usage{
konp_test(time, delta, trt, n.perm, n.impu = 1)
}
\arguments{
\item{time}{The follow-up time.}

\item{delta}{A binary status vector, where 0 stands for censored observation and 1 for events.}

\item{trt}{Group vector, must contain at least two unique values.}

\item{n.perm}{The number of permutations.}

\item{n.impu}{The number of imputations, for each imputation n.perm permutations will be executed.}
}
\value{
Two test statistics and their respective P-values are returned: \cr 

\code{pv_chisq} - returns the P-value based on the Chi-square statistic. \cr 
\code{pv_lr} - returns the P-value based on the likelihood ratio statistic. \cr 
\code{chisq_test_stat} - returns the Chi-square test statistic. \cr 
\code{lr_test_stat} - returns the likelihood ratio test statistic.
}
\description{
KONP Tests for Equality of K Distributions for Right-Censored Data
}
\details{
The KONP test is a powerful non-parametric K-sample survival test, for equality of K distributions.
 It is consistent against any form of difference between the two groups,
 and has been shown to offer more power than alternative tests, in non-proportional hazards data.
}
\examples{
## Generate some data to preform the test
set.seed(1)
n <- 50
time <- rexp(n)
delta <- sample(c(0,1),n,TRUE)
trt <- c(rep(1,25),rep(2,25))

konp_test(time,delta,trt,n.perm=10^3)

}
