% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_2_sample.R
\name{sample_group}
\alias{sample_group}
\title{Create permutation for group vector}
\usage{
sample_group(group, status_orginal, imputed_altern_status_vec)
}
\arguments{
\item{group}{A binary vector indicating group.}

\item{status_orginal}{A binary status vector for the original data, 0 for censored observations and 1 for events.}

\item{imputed_altern_status_vec}{A binary imputed status vector for observations that will change group in the permuted data.}
}
\value{
A permuted vector of group such that there are at least two events in each group .
}
\description{
Create permutation for group vector
}
\details{
This is an inner function inside KONPSURV package, not for the users.
}
