% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_K_sample.R
\name{sample_group_K_sample}
\alias{sample_group_K_sample}
\title{Create permutation for group vector with K values}
\usage{
sample_group_K_sample(group, imputed_status_matrix)
}
\arguments{
\item{group}{A binary vector indicating group, contains values from 1 to K}

\item{imputed_status_matrix}{A matrix contianing the status each obsevations will recieve for each group it belongs.}
}
\value{
A permuted vector of group such that there are at least two events in each group .
}
\description{
Create permutation for group vector with K values
}
\details{
This is an inner function inside KONPsurv package, not for the users.
}
