% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_2_sample.R
\name{sample_trt}
\alias{sample_trt}
\title{Create permutation for group vector}
\usage{
sample_trt(trt, delta_orginal, imputed_altern_delta_vec)
}
\arguments{
\item{trt}{A binary vector indicating group.}

\item{delta_orginal}{A binary status vector for the original data, 0 for censored observations and 1 for events.}

\item{imputed_altern_delta_vec}{A binary imputed status vector for observations that will change group in the permuted data.}
}
\value{
A permuted vector of trt such that there are at least two events in each group .
}
\description{
Create permutation for group vector
}
\details{
This is an inner function inside KONPSURV package, not for the users.
}
