% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHGKM_K_sample.R
\name{sample_trt_K_sample}
\alias{sample_trt_K_sample}
\title{Create permutation for group vector with K values}
\usage{
sample_trt_K_sample(trt, imputed_delta_matrix)
}
\arguments{
\item{trt}{A binary vector indicating group, contains values from 1 to K}

\item{imputed_delta_matrix}{A matrix contianing the delta each obsevations will recieve for each group it belongs.}
}
\value{
A permuted vector of trt such that there are at least two events in each group .
}
\description{
Create permutation for group vector with K values
}
\details{
This is an inner function inside KONPSURV package, not for the users.
}
